feat(tower): add HTTP client layer with server/client modules
Reorganize the crate into http::server and http::client modules (with
shared extractors and common helpers) and add a new client layer that
produces SpanKind::Client spans, records http.client.* metrics, and
injects the current trace context into outgoing request headers so
downstream servers continue the trace.

Layer selection is now controlled by http-server/http-client cargo
features (both on by default), with grpc-server/grpc-client reserved for
the upcoming gRPC layers. This is a breaking change: HTTPLayer and
HTTPLayerBuilder are replaced by http::server::Layer and
http::server::LayerBuilder.
