+++ opentelemetry-instrumentation-tower/src/common/mod.rs
@@ -1,0 +1,6 @@
+//! Building blocks shared by the HTTP (and future gRPC) server and client layers.
+
+pub(crate) mod attributes;
+pub(crate) mod instrumentation;
+pub(crate) mod propagation;
+pub(crate) mod status;
